# doc/apidoc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Khazana POS API",
    "description": "## Khazana POS Backend API\n\nAvailable modules:\n- **Auth** – Login, profile, change password\n- **Users** – User management and role assignment\n- **Sales** – Cash, credit, VAT cash, VAT credit sales\n- **Inventory** – Items, stock receive / issue / adjust\n- **Customers** – Customer management, ledger, payments\n- **Roles** – Role management\n- **Menus** – Menu items\n- **Permissions** – Role-based permissions\n- **Orders** – Sales orders (regular and VAT)\n- **Pricing** – Sale prices and cost prices\n- **Finance** – Money receive and cash purchases\n- **Packets** – Packet stock management\n- **NC Adjustment** – Non-conformance adjustments\n- **Assortment** – Assortment records\n- **Reports** – Sales, stock, item-wise, customer statements\n- **Admin** – Branches, settings, banks, audit log\n- **Dashboard** – Summary statistics\n- **Categories** – Item category management",
    "version": "1.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "required": ["userName", "password"],
        "properties": {
          "userName": { "type": "string", "example": "admin" },
          "password": { "type": "string", "example": "password123" }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "required": ["currentPassword", "newPassword"],
        "properties": {
          "currentPassword": { "type": "string", "example": "oldpass" },
          "newPassword": { "type": "string", "example": "newpass" }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": { "type": "integer", "example": 100 },
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 10 },
          "totalPages": { "type": "integer", "example": 10 },
          "hasNextPage": { "type": "boolean", "example": true },
          "hasPreviousPage": { "type": "boolean", "example": false }
        }
      },
      "CategoryDto": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "code": { "type": "string", "example": "SWEETS" },
          "name": { "type": "string", "example": "Sweets" },
          "remarks": { "type": "string", "example": "All sweet items" }
        }
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login and receive JWT token",
        "operationId": "AuthController_login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns access token and user profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": { "type": "string" },
                    "user": { "type": "object" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/profile": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current user profile",
        "operationId": "AuthController_getProfile",
        "security": [{ "access-token": [] }],
        "responses": {
          "200": { "description": "Returns authenticated user profile" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/auth/change-password": {
      "patch": {
        "tags": ["Auth"],
        "summary": "Change password for current user",
        "operationId": "AuthController_changePassword",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePasswordDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Password changed successfully" },
          "400": { "description": "Current password is incorrect" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/auth/verify-email": {
      "post": {
        "tags": ["Auth"],
        "summary": "Verify email address using token from verification link",
        "operationId": "AuthController_verifyEmail",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token"],
                "properties": { "token": { "type": "string", "example": "a3f8b2c1d4e5..." } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Email verified successfully" },
          "400": { "description": "Invalid or expired token" }
        }
      }
    },
    "/auth/resend-verification-email": {
      "post": {
        "tags": ["Auth"],
        "summary": "Resend email verification link",
        "operationId": "AuthController_resendVerificationEmail",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": { "email": { "type": "string", "format": "email", "example": "user@example.com" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Verification link sent if email is registered" }
        }
      }
    },
    "/auth/send-verification-email": {
      "post": {
        "tags": ["Auth"],
        "summary": "Send verification email (admin / manual trigger)",
        "operationId": "AuthController_sendVerificationEmail",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": { "email": { "type": "string", "format": "email", "example": "user@example.com" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Verification link sent if email is registered" }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Request a 6-digit password reset code sent to email",
        "operationId": "AuthController_forgotPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": { "email": { "type": "string", "format": "email", "example": "user@example.com" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Reset code sent if email is registered (always 200 to prevent enumeration)" }
        }
      }
    },
    "/auth/verify-reset-code": {
      "post": {
        "tags": ["Auth"],
        "summary": "Validate 6-digit password reset code from email",
        "operationId": "AuthController_verifyResetCode",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "code"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "user@example.com" },
                  "code": { "type": "string", "minLength": 6, "maxLength": 6, "example": "847291" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Code verified — user can now call reset-password" },
          "400": { "description": "Invalid or expired code" }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Set a new password after verifying reset code",
        "operationId": "AuthController_resetPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "code", "newPassword"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "user@example.com" },
                  "code": { "type": "string", "minLength": 6, "maxLength": 6, "example": "847291" },
                  "newPassword": { "type": "string", "minLength": 8, "example": "NewSecure@123", "description": "Min 8 chars, must contain uppercase, lowercase, and a number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Password reset successfully" },
          "400": { "description": "Invalid code, expired session, or password too weak" }
        }
      }
    },
...
    "/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get all item categories",
        "operationId": "CategoriesController_findAll",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "default": 1, "minimum": 1 }, "description": "Page number" },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 100 }, "description": "Items per page" }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Category retrieved successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": { "type": "array", "items": { "type": "object" } },
                        "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Create a new category",
        "operationId": "CategoriesController_create",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoryDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Category created successfully" },
          "409": { "description": "Category code already exists" }
        }
      }
    },
    "/categories/{id}": {
      "patch": {
        "tags": ["Categories"],
        "summary": "Update category by ID",
        "operationId": "CategoriesController_update",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Category UUID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object" } }
          }
        },
        "responses": {
          "200": { "description": "Category updated successfully" }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "Delete category by ID",
        "operationId": "CategoriesController_remove",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Category UUID" }
        ],
        "responses": {
          "200": { "description": "Category deleted successfully" },
          "404": { "description": "Category not found" }
        }
      }
    }
  },
...
}